Database, controllers, services, guard, decorator,
Wrote a controller and a user service. Added authentication via cookies. Wrote auth guard. Connected the database. Created all entities for the database. Connected env. Wrote a decorator to retrieve the user. Wrote a controller to search for articles with queries and create an asset (not tested yet). Wrote a controller and service for tag. Wrote a controller for subscriptions to other users (haven't tested it yet). And a lot of other things.